// powerapps-project-planner-20260612/src/generated/models/Vibe_decisionlogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_decisionlogsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_decisionlogsstatecode = keyof typeof Vibe_decisionlogsstatecode;
export const Vibe_decisionlogsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_decisionlogsstatuscode = keyof typeof Vibe_decisionlogsstatuscode;

export interface Vibe_decisionlogsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_decisionlogsstatecode;
  statuscode?: Vibe_decisionlogsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_date?: string;
  vibe_decisionlogid: string;
  vibe_description?: string;
  "vibe_madebyid@odata.bind"?: string;
  "vibe_phaseid@odata.bind"?: string;
  "vibe_projectid@odata.bind"?: string;
  vibe_rationale?: string;
  vibe_title?: string;
}

export interface Vibe_decisionlogs extends Vibe_decisionlogsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_madebyidname?: string;
  vibe_phaseidname?: string;
  vibe_projectidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_madebyid?: object;
  _vibe_madebyid_value?: string;
  vibe_phaseid?: object;
  _vibe_phaseid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}