// powerapps-project-planner-20260612/src/generated/models/Vibe_meetingnotesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_meetingnotesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_meetingnotesstatecode = keyof typeof Vibe_meetingnotesstatecode;
export const Vibe_meetingnotesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_meetingnotesstatuscode = keyof typeof Vibe_meetingnotesstatuscode;

export interface Vibe_meetingnotesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_meetingnotesstatecode;
  statuscode?: Vibe_meetingnotesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_date?: string;
  vibe_meetingnoteid: string;
  vibe_notes?: string;
  "vibe_phaseid@odata.bind"?: string;
  "vibe_projectid@odata.bind"?: string;
  vibe_title?: string;
}

export interface Vibe_meetingnotes extends Vibe_meetingnotesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_phaseidname?: string;
  vibe_projectidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_phaseid?: object;
  _vibe_phaseid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}