// powerapps-project-planner-20260612/src/generated/models/Vibe_activityfeedentriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_activityfeedentriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_activityfeedentriesstatecode = keyof typeof Vibe_activityfeedentriesstatecode;
export const Vibe_activityfeedentriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_activityfeedentriesstatuscode = keyof typeof Vibe_activityfeedentriesstatuscode;
export const Vibe_activityfeedentriesvibe_eventtype = {
  100000000: 'STATUS_CHANGE',
  100000001: 'TEAM_CHANGE',
  100000002: 'PHASE_COMPLETED',
  100000003: 'BUDGET_THRESHOLD',
  100000004: 'RISK_CREATED',
  100000005: 'MILESTONE_REACHED',
  100000006: 'DELIVERABLE_SUBMITTED',
  100000007: 'DECISION_MADE'
} as const;
export type Vibe_activityfeedentriesvibe_eventtype = keyof typeof Vibe_activityfeedentriesvibe_eventtype;

export interface Vibe_activityfeedentriesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_activityfeedentriesstatecode;
  statuscode?: Vibe_activityfeedentriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_activityfeedentryid: string;
  "vibe_actorid@odata.bind"?: string;
  vibe_description?: string;
  vibe_eventtype?: Vibe_activityfeedentriesvibe_eventtype;
  vibe_name?: string;
  "vibe_projectid@odata.bind"?: string;
  vibe_relatedrecordurl?: string;
  vibe_timestamp?: string;
}

export interface Vibe_activityfeedentries extends Vibe_activityfeedentriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_actoridname?: string;
  vibe_actoridyominame?: string;
  vibe_eventtypename?: string;
  vibe_projectidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_actorid?: object;
  _vibe_actorid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}