// powerapps-project-planner-20260612/src/generated/models/Vibe_projectteammembersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_projectteammembersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_projectteammembersstatecode = keyof typeof Vibe_projectteammembersstatecode;
export const Vibe_projectteammembersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_projectteammembersstatuscode = keyof typeof Vibe_projectteammembersstatuscode;

export interface Vibe_projectteammembersBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_projectteammembersstatecode;
  statuscode?: Vibe_projectteammembersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_allocationpercentage?: number;
  "vibe_contactid@odata.bind"?: string;
  vibe_enddate?: string;
  vibe_hourlyrate?: number;
  vibe_isactive?: boolean;
  vibe_name?: string;
  "vibe_projectid@odata.bind"?: string;
  vibe_projectteammemberid: string;
  "vibe_roleid@odata.bind"?: string;
  vibe_startdate?: string;
  "vibe_userid@odata.bind"?: string;
}

export interface Vibe_projectteammembers extends Vibe_projectteammembersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_contactidname?: string;
  vibe_contactidyominame?: string;
  vibe_isactivename?: string;
  vibe_projectidname?: string;
  vibe_roleidname?: string;
  vibe_useridname?: string;
  vibe_useridyominame?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_contactid?: object;
  _vibe_contactid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
  vibe_roleid?: object;
  _vibe_roleid_value?: string;
  vibe_userid?: object;
  _vibe_userid_value?: string;
}